Use Domain Events in place of ZF2 Events to trigger business logic and introduce Organization as the issuer entity of an account
Each command generates one or more domain event and each domain event
could trigger additional business logic, such as notifications or
interaction with external systems and so on.
The business logic that is listening to events to happen is now
listening on domain events triggered by Event Store and not more
duplicated ZF2 copies.

Each personal account is now associated with the an organization.
It tracks all transactions related to the owner respect that orga.

The domain object getId method now returns a string while getUuid
returns a Uuid. A domain object has now a string representation: its
string id